+++ src/main/java/com/thoughtworks/capacity/gtb/mvc/Repository/UserRepositoryImpl.java
@@ -13,19 +13,19 @@ import java.util.Map;
 @Repository
 public class UserRepositoryImpl implements UserRepository{
 
-    private final Map<Integer,User> users = new HashMap<>();
+    private final Map<Long,User> users = new HashMap<>();
 
     private final List<Education> educations = new ArrayList<>();
 
     @Override
-    public Integer save(User user) {
+    public long save(User user) {
         user.setId(users.size()+1);
         users.put(user.getId(),user);
         return user.getId();
     }
 
     @Override
-    public User findById(Integer id) {
+    public User findById(long id) {
         return users.get(id);
     }
 
@@ -35,7 +35,7 @@ public class UserRepositoryImpl implements UserRepository{
     };
 
     @Override
-    public List<Education> findEducationByid(Integer id) {
+    public List<Education> findEducationByid(long id) {
         List<Education> reEducations = new ArrayList<>();
         educations.forEach(education -> {
             if(education.getUserId() == id) {
